Add import Numpy  library
numpy is a library which makes it easier to use advanced mathematical formulas in python such as linear algebra, Fourier transform, and random number capabilities

After we have imported our libraries we need to add a function, this is a sigmoid function, which is a type of non-linearity that we have chosen for this neural network. It is not the only type of non-linearity that can be chosen, but is has simple analytical features and is easy to use and understand. In practice, large-scale deep learning systems use piecewise-linear functions because they are much less expensive to evaluate.
The implementation of this function does double duty. If the deriv=True flag is passed in, the function instead calculates the derivative of the function, which is used in the error back propagation step. I mentioned back propagation earlier in the tutorial so if you don’t know what this is then please refer back to the How Does A Neural Network Learn? section.
def nonlin(x, deriv=False):
  if(deriv==True):
     return (x*(1-x))
  return 1/(1+np.exp(-x))
When we have added this function we need to import our data as an array. This creates the input matrix.
X = np.array([[0,0,1], 
      [0,1,1], 
      [1,0,1],
      [1,1,1]])
When we have inputted the data that we want to train the neural network with we need to add the output data. This is where the feedback happens because we are telling the program what we want the output to be so it can match the input to the desired output and if the current output is wrong we can tell the program and then correct it by putting it back through the neurons and this is where it starts to learn.
y = np.array([[0], 
        [1], 
        [1], 
        [0]])
The seed for the random generator is set so that it will return the same random numbers each time.
np.random.seed(1)
Then we need to create the neurons. Then we initialise the weights to random values. syn0 are the weights between the input layer and the hidden layer. It is a 3x4 matrix because there are two input weights plus a bias term (=3) and four nodes in the hidden layer (=4). syn1 are the weights between the hidden layer and the output layer. It is a 4x1 matrix because there are 4 nodes in the hidden layer and one output. Note that there is no bias term feeding the output layer in this example. The weights are initially generated randomly because optimisation tends not to work well when all the weights start at the same value.
syn0 = 2*np.random.random((3,4)) - 1
syn1 = 2*np.random.random((4,1)) - 1