Choosing a task by “Exploitation” or “ Exploration” algorithm
Algorithm 1: reference
A Two-stage Framework and Reinforcement
Learning-based Optimization Algorithms for
Complex Scheduling Problems

2103.05847v1 reference 

In the algorithm below, threshold value ε is the probability
of choosing an action randomly (line 3 to 4). Otherwise, the
action is taken according to Q. The Q values of all actions at
state st are calculated by the Q-network, which is trained by
Eq. (16). Only the actions in A(St) are considered at time step
t, so we record the Q value of all available actions Q in list,
and then get the action with the highest Q value. see line 7 to
10.
After training, the value function (Q-network) has been
finalized. In the testing process, the algorithm always select
the action with the highest Q value for seeking the satisfactory
solution.
B. OR algorithms in the rear stage
Once the solution of the prior stage has been determined,
there are several types of algorithms for dealing with the
problem in the rear stage. Dynamic programming and a
constructive heuristic algorithm are applied in this study.
1) Constructive heuristic algorithm: A constructive heuris-
tic algorithm based on the density of residual tasks, named
HADRT[38], is introduced as one of the algorithms for se-
quencing and timing problem. The core of HADRT are the
expressions below:
f(i) = g(i) + h(i) = X
i|ri>0
1 +
ET − wei
dl
(17)
k = arg max f(i) (18)
where, g(i) is the number of scheduled tasks after adding the
i-th task; h(i) equals the ratio between the remaining time
length and the average duration of tasks.
Under the premise of satisfying all constraints, execution
start time for each scheduled task is determined as follows:
esk = max(wbk, eek−1 + ctk,k−1) (19)
where ctk,k−1 indicates the preparation time required from the
end of task k − 1 to the start of task k.
The expectation of total profit in the scheme obtained by
HADRT is maximized, under the following conditions:
• Values of preparation time and durations of tasks are fixed
or follow a determined distribution;
• The profit of each task is a constant;
• The ETW of any task is not completely covered by the
ETW of another task.
The above statement has been proven by mathematical
induction and reduction. The detailed process of the proof is
derived in [38], which is not discussed here. The algorithm is
organized as the sequence of the following steps:
Step 1: Initialization: the latest end time of the scenario,
the set of candidate tasks;
Step 2: Calculate f(i) for each available task;
Step 3: Find the task k by Eq. (17) and (18);
Step 4: Set the execution start time of task k by Eq. (19);
Step 5: Update the latest end time of the scenario if the
task k is scheduled;
Step 6: Repeat Step 2 to Step 5 until all of the tasks are
considered.
2) Dynamic programming: Dynamic programming (DP) is
one of the commonly used methods for solving combinatorial
optimization problems. DP guarantees the optimal solution for
the timing problem in the rear stage with time complexity
O(n
2
), under the assumptions of: 1) The sequence of tasks is
fixed. 2) Time could be discretized. The above statement can
be proved because the problem meets the optimality principle
of dynamic programming