Algorithm 2: Dynamic programming
It is worth noting that this process meets the principle of
optimality and without aftereffect[39]. As shown in Figure 3,
B[i][t] is the maximum total profit that the top i-th tasks can
get in the time t, which is calculated as
B[i][t] = max
k
{B[i − 1][t], B[i − k] [t − di − cti,k] + pi}
(20)
Based on the above, the pseudo-code of solving the timing
problem by DP is as follows:
Algorithm 2: Dynamic programming
Input: A sequence of tasks T
Output: Objective function value of the optimal
solution B[|T|][M axT ime]
1 Parameters initialization
2 M axT ime ← timespan
3 foreach i ∈ T do
4 for t = 0 : M axT ime do
5 for k = 0 : i do
6 T empB ← B[i − k][t − di − cti,k] + pi
7 if satisfy all constraints and
T empB > B[i][t] then
8 B[i][t] ← T empB
9 else
10 B[i][t] ← B[i − 1][t]
3) Summary on HADRT and DP: HADRT and DP are
representative heuristic algorithm and mathematical program-
ming algorithm in traditional OR area, respectively. Reward
of time step t is calculated by the result of HADRT or
DP at time step t, therefore, the performance of HADRT
and DP affects the rewards of RL. The range of rewards
obtained by HADRT should be contained in the interval
h
− maxi,j p
j
i
, maxi,j p
j
i
i
, in which p
j
i
is the profit of task
i on resource j. Correspondingly, rewards obtained by DP
should be non-negative because DP guarantees the optimality
of timing problem, and the range of rewards obtained by DP
is reduced to h
0, maxi,j p
j
i
i
.
Both the time complexity of HADRT and DP are O(n
2
), in
which, n represents the scale of the problem. Therefore, the
time complexity of these two algorithms is in a low level.
We discuss the performance and consumption of these two
approaches in the experiments.